Move zlib frame decompressor out of nexrad.
It's generally useful for more NOAAPORT products.
